requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
coverage==5.5
deprecation==2.1.0
eyed3==0.9.6
filetype==1.0.7
idna==2.10
packaging==20.9
pyparsing==2.4.7
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
youtube-dl==2021.6.6
youtube-search==2.1.0
